chore: move utility function into its' own file
